some change logic with add information but i will think that need some rebuld
