+++ src/screens/ScreenFilms/ScreenFilms.styles.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components/native';
+import {typography, colors} from '../../styles/theme';
+import {Dimensions, SafeAreaView} from 'react-native';
+
+export const Container = styled.SafeAreaView``;
+export const Block = styled.View``;
+export const TextStyled = styled.Text``;
+export const StyledFlatList = styled.FlatList``;
